Moved Tag1 and Tag2 to a different package
