+++ library/src/test/java/com/matthewtamlin/java_compiler_utilities/element_supplier/annotated_element_supplier/Tag1.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.java_compiler_utilities;
+package com.matthewtamlin.java_compiler_utilities.element_supplier.annotated_element_supplier;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
